homework/Class 2 HomeWork/bonus/Task 3/requirements.txt
yt-dlp>=2023.12.30
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0